% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-class.R
\name{as_omics_braid_data}
\alias{as_omics_braid_data}
\title{Coerce supported objects to OmicsBraid data}
\usage{
as_omics_braid_data(x, ...)
}
\arguments{
\item{x}{Object to coerce.}

\item{...}{Additional arguments.}
}
\description{
Coerce supported objects to OmicsBraid data
}
