% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_braid_heatmap}
\alias{plot_braid_heatmap}
\title{Plot a braid heatmap across entities and omics}
\usage{
plot_braid_heatmap(
  result,
  entities = NULL,
  omic_order = result$settings$omic_order
)
}
\arguments{
\item{result}{An `omics_braid_result`.}

\item{entities}{Optional entities; default top 25 by integrated adjusted p-value.}

\item{omic_order}{Optional omic order.}
}
\value{
A ggplot object.
}
\description{
Plot a braid heatmap across entities and omics
}
