% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_evidence_forest}
\alias{plot_evidence_forest}
\title{Plot an Omics Evidence Forest}
\usage{
plot_evidence_forest(result, entity, omic_order = result$settings$omic_order)
}
\arguments{
\item{result}{An `omics_braid_result`.}

\item{entity}{Entity/pathway to display.}

\item{omic_order}{Optional omic order.}
}
\value{
A ggplot object.
}
\description{
Plot an Omics Evidence Forest
}
