% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-class.R
\name{validate_omics_braid_data}
\alias{validate_omics_braid_data}
\title{Validate an OmicsBraid data object}
\usage{
validate_omics_braid_data(x)
}
\arguments{
\item{x}{An `omics_braid_data` object.}
}
\value{
Invisibly returns `TRUE` or throws an informative error.
}
\description{
Validate an OmicsBraid data object
}
